Create 4ujXDK9mpK3tvz4PnXsopQC53PRLxpKRC4mkkogUmS9q
WUSD is a test and experimental token created on the Solana blockchain for educational and testing purposes. This token is designed to simulate a stablecoin environment, inspired by popular tokens like USDT. WUSD can be used in decentralized exchanges (DEXs), wallet integration tests, and other blockchain-related experiments. Please note that this token holds no real-world value and is intended solely for testing on the Solana network.